failure to set output trigger (#77)
closes #76 

When setting up output triggering, DCAM_IDPROP_OUTPUTTRIGGER_SOURCE was
being set for all trigger kinds but it's not always writable. The
"trigger source" is only really for when using
DCAMPROP_OUTPUTTRIGGER_KIND__PROGRAMABLE.

Adds a test that runs through setting up an output trigger on line 1
configured to output when exposure starts.