Fix setting ROI (#86)
closes #85 

adds test, updates changelog

Changes the ROI setting procedure so it's more robust:
1. Set offset to 0, so size can have full range of sensor.
2. Set the size
3. Set the offset

Sometimes the offset has to be divisible by 4. When setting a parameter
that is not appropriately divisible, the call just fails with an invalid
parameter which is not very helpful.

Fortunately, when running in "progressive" mode (the default for
light-sheet), there are fewer restrictions. I'm not exactly sure how to
query the restrictions, so the approach for this PR is to make sure
we're setting the mode early.