execute_trigger forces a software trigger in all modes (#195)
closes #194 
depends on acquire-project/acquire-common#34

Changes the `execute_trigger` behavior to force a software trigger when
called regardless of the camera configuration. This helps with
`acquire_abort()`.

Interestingly, we seem to already have an `abort-finite-acquisition`
test. I wonder if it can be improved to cover the problem Micah ran
into. May need to just add a check that the everything is restartable
after the abort.

---------

Co-authored-by: Alan Liddell <aliddell@chanzuckerberg.com>